Move common methods to new aioseo helper class
