+++ src/actions/importing/aioseo/aioseo-cleanup-action.php
@@ -6,7 +6,6 @@ namespace Yoast\WP\SEO\Actions\Importing\Aioseo;
 use wpdb;
 use Yoast\WP\SEO\Actions\Importing\Abstract_Aioseo_Importing_Action;
 use Yoast\WP\SEO\Helpers\Options_Helper;
-use Yoast\WP\SEO\Helpers\Wpdb_Helper;
 
 /**
  * Importing action for cleaning up AIOSEO data.
@@ -47,36 +46,17 @@ class Aioseo_Cleanup_Action extends Abstract_Aioseo_Importing_Action {
 	protected $wpdb;
 
 	/**
-	 * The wpdb helper.
-	 *
-	 * @var Wpdb_Helper
-	 */
-	protected $wpdb_helper;
-
-	/**
 	 * Class constructor.
 	 *
 	 * @param wpdb           $wpdb        The WordPress database instance.
 	 * @param Options_Helper $options     The options helper.
-	 * @param Wpdb_Helper    $wpdb_helper The wpdb_helper helper.
 	 */
 	public function __construct(
 		wpdb $wpdb,
-		Options_Helper $options,
-		Wpdb_Helper $wpdb_helper
+		Options_Helper $options
 	) {
-		$this->wpdb        = $wpdb;
-		$this->options     = $options;
-		$this->wpdb_helper = $wpdb_helper;
-	}
-
-	/**
-	 * Retrieves the AIOSEO table name along with the db prefix.
-	 *
-	 * @return string The AIOSEO table name along with the db prefix.
-	 */
-	protected function get_aioseo_table() {
-		return $this->wpdb->prefix . 'aioseo_posts';
+		$this->wpdb    = $wpdb;
+		$this->options = $options;
 	}
 
 	/**
@@ -89,21 +69,12 @@ class Aioseo_Cleanup_Action extends Abstract_Aioseo_Importing_Action {
 	}
 
 	/**
-	 * Determines if the AIOSEO database table exists.
-	 *
-	 * @return bool True if the table is found.
-	 */
-	protected function aioseo_exists() {
-		return $this->wpdb_helper->table_exists( $this->get_aioseo_table() ) === true;
-	}
-
-	/**
 	 * Just checks if the cleanup has been completed in the past.
 	 *
 	 * @return int The total number of unimported objects.
 	 */
 	public function get_total_unindexed() {
-		if ( ! $this->aioseo_exists() ) {
+		if ( ! $this->aioseo_helper->aioseo_exists() ) {
 			return 0;
 		}
 
@@ -118,7 +89,7 @@ class Aioseo_Cleanup_Action extends Abstract_Aioseo_Importing_Action {
 	 * @return int|false The limited number of unindexed posts. False if the query fails.
 	 */
 	public function get_limited_unindexed_count( $limit ) {
-		if ( ! $this->aioseo_exists() ) {
+		if ( ! $this->aioseo_helper->aioseo_exists() ) {
 			return 0;
 		}
 
@@ -176,11 +147,11 @@ class Aioseo_Cleanup_Action extends Abstract_Aioseo_Importing_Action {
 	 * @return string The query to use for importing or counting the number of items to import.
 	 */
 	public function truncate_query() {
-		if ( ! $this->aioseo_exists() ) {
+		if ( ! $this->aioseo_helper->aioseo_exists() ) {
 			return true;
 		}
 
-		$table = $this->get_aioseo_table();
+		$table = $this->aioseo_helper->get_table();
 
 		return "TRUNCATE TABLE {$table}";
 	}
